backend/requirements.txt
meson==1.6.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Marshmallow==3.23.0
marshmallow-sqlalchemy==1.1.0
python-dotenv==1.0.1
flask-cors==5.0.0
pytest==8.3.3
pytest-flask==1.3.0